add confirmation on delete

+++ src/main/java/org/example/service/UpdateHandler.java
@@ -674,22 +674,35 @@ public class UpdateHandler {
 
     public void removeRecording(CallbackQuery callbackQuery) throws TelegramApiException
     {
-        SendMessage sendMessage = new SendMessage();
-        sendMessage.setChatId(callbackQuery.getMessage().getChatId());
-
-        String messageId = callbackQuery.getData().substring("remove_".length());
-
-        Long userId = callbackQuery.getFrom().getId();
-        logger.debug("Removing message by id: {}, user: {}", messageId, userId);
-        System.out.println("Removing message by id: " + messageId + ", user: " + userId);
-        int updatedRows = userService.removeRecordByUserIdAndMessageId(userId, messageId);
-        sendMessage.setReplyToMessageId(Integer.valueOf(messageId));
-        if (updatedRows > 0) {
-            sendMessage.setText(OK_REMOVED);
-        } else {
-            sendMessage.setText(RECORDING_NOT_FOUND);
+        if (callbackQuery.getData().contains("confirm"))
+        {
+            confirmRemovingRecording(callbackQuery);
         }
-        executeFunction.execute(sendMessage);
+        else
+        {
+            SendMessage sendMessage = new SendMessage();
+            sendMessage.setChatId(callbackQuery.getMessage().getChatId());
+
+            if (callbackQuery.getData().contains("no"))
+            {
+                sendMessage.setText("OK");
+            }
+            else {
+                String messageId = callbackQuery.getData().substring("remove_".length());
+
+                Long userId = callbackQuery.getFrom().getId();
+                logger.debug("Removing message by id: {}, user: {}", messageId, userId);
+                System.out.println("Removing message by id: " + messageId + ", user: " + userId);
+                int updatedRows = userService.removeRecordByUserIdAndMessageId(userId, messageId);
+                sendMessage.setReplyToMessageId(Integer.valueOf(messageId));
+                if (updatedRows > 0) {
+                    sendMessage.setText(OK_REMOVED);
+                } else {
+                    sendMessage.setText(RECORDING_NOT_FOUND);
+                }
+            }
+            executeFunction.execute(sendMessage);
+        }
     }
 
     public void returnMainMenu(Message message) throws TelegramApiException
@@ -1002,6 +1015,17 @@ public class UpdateHandler {
         sendMessage.setEntities(Arrays.asList(messageEntity));
         sendMessage.setText("OK. You can go through the tutorial whenever you want using the /tutorial command");
         sendMessage.setReplyMarkup(buttonsService.getInitMenuButtons());
+        executeFunction.execute(sendMessage);
+    }
+
+    public void confirmRemovingRecording(CallbackQuery callbackQuery) throws TelegramApiException
+    {
+        String messageId = callbackQuery.getData().substring("confirmremove_".length());
+
+        SendMessage sendMessage = new SendMessage();
+        sendMessage.setChatId(callbackQuery.getMessage().getChatId());
+        sendMessage.setText("Do you really want to delete the recording?");
+        sendMessage.setReplyMarkup(ButtonsService.getRemovingConfirmationButtons(Integer.valueOf(messageId)));
         executeFunction.execute(sendMessage);
     }
